// src/types/editor.ts
export type ImageSettings = {
    brightness: number;
    exposure: number;
    contrast: number;
    saturation: number;

    shadows: number;
    highlights: number;
    gamma: number;
    blacks: number;
    whites: number;

    temperature: number;
    tint: number;

    sharpness: number;

    width: number;
    height: number;
};